Update components due to change with modal refactor
